agents/planning/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
batchgenerators==0.25.1
boto3==1.37.18
botocore==1.37.18
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dicom2nifti==2.5.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.2.0.post0
lightning-utilities==0.14.2
linecache2==1.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
mednextv1 @ git+https://github.com/MIC-DKFZ/MedNeXt.git@0b78ed869fbd1cc2fd38754d2f8519f1b72d43ba
MedPy==0.5.2
monai==1.4.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.2
pandas==2.2.0
pillow==11.1.0
propcache==0.3.0
pydicom==3.0.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0
requests==2.32.3
s3transfer==0.11.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.11.4
SimpleITK==2.3.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.1.2
torchmetrics==1.7.0
tqdm==4.67.1
traceback2==1.4.0
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
unittest2==1.1.0
urllib3==2.3.0
yarl==1.18.3
